requirements.txt
numpy
pillow
torch
torchvision
hnswlib
tqdm
pyyaml
scikit-learn
matplotlib
